Squashed commit of the following:

commit 1c7cfc051275c991b96878fdfde0b8851be789eb
Author: Michael Altizer <mialtize@cisco.com>
Date:   Tue Mar 20 18:16:47 2018 -0400

    ips_byte_math: Fix UBSAN left shift of negative value runtime error

commit 849a83586f85a85c8192e7e48aac07cf09e330b5
Author: Michael Altizer <mialtize@cisco.com>
Date:   Tue Mar 20 18:16:26 2018 -0400

    ips_byte_jump: Fix UBSAN left shift of negative value runtime error

commit 51f5501044cd978079c4e41d45cce5e522e3063e
Author: Michael Altizer <mialtize@cisco.com>
Date:   Tue Mar 20 17:49:14 2018 -0400

    hashfcn: Fix UBSAN left shift of negative value runtime error

commit a5cbf457bcdfcdbb532bc36aa353c6baea8241bb
Author: Michael Altizer <mialtize@cisco.com>
Date:   Tue Mar 20 17:07:20 2018 -0400

    parameter: Fix UBSAN shift exponent is too large for 32-bit type runtime error

commit 015abe650696b0be1e18a429e62efc096d228976
Author: Michael Altizer <mialtize@cisco.com>
Date:   Tue Mar 20 16:49:10 2018 -0400

    binder: Fix UBSAN invalid value type runtime error

commit e80d3ffca94a8b508a4e608bc2d90df7a18258cb
Author: Michael Altizer <mialtize@cisco.com>
Date:   Tue Mar 20 16:34:34 2018 -0400

    wizard: Fix UBSAN out-of-bounds access runtime error

commit 540a3cd4973db76f08e12e316727ea2d63d71aad
Author: Michael Altizer <mialtize@cisco.com>
Date:   Tue Mar 20 16:33:35 2018 -0400

    hashfcn: Fix UBSAN integer overflow runtime error

commit 470dc447d63533a784dbfa85e64a39227faffaba
Author: Michael Altizer <mialtize@cisco.com>
Date:   Tue Mar 20 19:22:16 2018 -0400

    extra: Port some CMake options from Snort prime

commit 2dbd95924b68d78104302ec6b62d8ee3e0178760
Author: Michael Altizer <mialtize@cisco.com>
Date:   Tue Mar 20 15:13:50 2018 -0400

    build: Add --enable-ub-sanitizer option for undefined behavior sanitizer

+++ src/framework/parameter.cc
@@ -253,7 +253,7 @@ static bool valid_multi(Value& v, const char* r)
     vector<string> list;
     split(s, list);
 
-    unsigned mask = 0;
+    unsigned long long mask = 0;
 
     for ( const auto& p : list )
     {
@@ -264,7 +264,7 @@ static bool valid_multi(Value& v, const char* r)
         unsigned idx = get_index(r, t);
 
         if ( idx < Value::mask_bits )
-            mask |= (1 << idx);
+            mask |= (1ULL << idx);
     }
     v.set_aux(mask);
     return true;
